% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2csv.R
\name{fit2csv}
\alias{fit2csv}
\title{Fit to CSV conversion}
\usage{
fit2csv(file, verbose = FALSE)
}
\arguments{
\item{file}{Fit file to convert}

\item{verbose}{print diagnostic messages}
}
\value{
A character filename of CSV
}
\description{
Fit to CSV conversion
}
\examples{
file = system.file("extdata", "example.fit",
package = "fit2csv")
res = fit2csv(file)
}
